requirements.txt
pygame-ce
PyOpenGL
PyOpenGL_accelerate
numpy